requirements.txt
# requirements.txt
upstage>=0.1
python-dotenv>=1.0.1
faiss-cpu>=1.8.0.post1
numpy>=1.26.4
requests>=2.32.3
beautifulsoup4>=4.12.3